requirements.txt
Flask==3.0.3
openai==1.61.0
python-dotenv==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.6